reset password funcionando

+++ src/controllers/ProductController.js
@@ -1,12 +1,6 @@
 const Product = require('../models/Product');
 const User = require('../models/User');
 
-  // função que verifica se o usuário é adm
-async function isADM(id){
-  const user = await User.findById(id)
-    if (user.role !== 1)
-      return res.status(203).send({ forbidden: 'you don\'t have authorization to do that' })
-}
 
 module.exports = {
   // traz todos os itens disponiveis do model de product
@@ -16,7 +10,11 @@ module.exports = {
   },
   // metodo para criação de itens novos, caso seja adm
   async store(req, res) {
-    isADM(req.userId)
+  // verifica se o usuário é adm
+    const user = await User.findById(req.userId)
+    if (!user.isAdmin)
+      return res.status(203).send({ forbidden: 'you don\'t have authorization to do that' })
+
     const product = await Product.find({ "title": req.body.title })
     if (product.length < 1) {
       const newItem = await Product.create(req.body)
@@ -31,7 +29,10 @@ module.exports = {
   },
   // atualização do produto existente através do id
   async update(req, res) {
-    isADM(req.userId);
+    const user = await User.findById(req.userId)
+    if (!user.isAdmin)
+      return res.status(203).send({ forbidden: 'you don\'t have authorization to do that' })
+
     const {quantity, price } = req.body;
 
     try {
@@ -40,10 +41,9 @@ module.exports = {
       return res.status(400).send({error: 'Quantidade não pode ser zero'})
       if(price < 1)
       return res.status(400).send({error: 'O preço não pode ser zerado'})
-      await product.update({
-        quantity: product.quantity += quantity,
-        price: product.price = price
-      })
+      product.quantity += quantity;
+      product.price = price
+      await product.save()
       return res.json(product)
     } catch (err) {
       return res.status(404).send({erro: 'Produto não encontrado'})
@@ -51,7 +51,9 @@ module.exports = {
   },
   // exclui o item do database
   async destroy(req, res) {
-    isADM(req.userId);
+    const user = await User.findById(req.userId)
+    if (!user.isAdmin)
+      return res.status(203).send({ forbidden: 'you don\'t have authorization to do that' })
 
     await Product.findByIdAndRemove(req.params.id)
     return res.status(200).send('Excluido com sucesso')
